Trying to fix ticket #16

+++ libdrqueue/semaphores.c
@@ -25,7 +25,7 @@
 #include <stdio.h>
 #include <errno.h>
 
-void semaphore_lock (int semid) {
+void semaphore_lock (int64_t semid) {
   struct sembuf op;
 
   op.sem_num = 0;
@@ -40,7 +40,7 @@ void semaphore_lock (int semid) {
   /*  fprintf (stderr,"Locked !!! semval: %i semid: %i\n",semctl (semid,0,GETVAL),semid); */
 }
 
-void semaphore_release (int semid) {
+void semaphore_release (int64_t semid) {
   struct sembuf op;
 
   /*  fprintf (stderr,"Unlocking... semval: %i semid: %i\n",semctl (semid,0,GETVAL),semid); */
